Rough implementation of difficulty settings.
I at least figured out a way to set the timer to different things. Haven't made any buttons or menus to actually change the difficulty, but I have found a way to set the timer to something different. Also would still need to find a way to hide things like the rotation buttons on easier difficulties.