Duplication only happens in editor.
The text duplication seems to only happen in the editor. When I generate a build and run that, it doesn't happen.  So, now I think the last thing to do is try the Photon CountdownTimer script one more time with this new knowledge, because our original timer method does not sync at all. It just now has a timer for both views.