Seems like I figured out how to disable buttons.
Right now it should only be disabling the left button, but when testing it it disabled all of them for some reason. Not really sure why that happened if I'm being honest. I'll still put in explicit lines to disable all the others, it just gives me cause for concern.