Work in progress on online code.
Trying to maybe rework it a bit, might help make it easier to manage everything. Ideally, if this works out like I think it should, it should allow me to keep management of everything that both players should be able to see, like the timer and quit + reset buttons, and the management of user input separate.