Timer updates.
Not perfect, but this at least makes it so the timer runs like it did before again. I was trying to get it to still run while only testing the Game scene, but that's having issues, so I'm putting that idea aside in favor of getting everything else working like the triggers.