Timer stops on both sides now.
Though, there's a chance they could be off by one second due to latency. Right now, I'm kind of thinking this is okay. It's not a huge deal that the timer be identical when the game is won. Also, the reset button is functional now, but it only works on one player's end. The game can be synced back up once the plug player moves, but ideally we want it to be synced from the beginning.