Reset button sort of works.
Basically, it only resets the scene on one player's end.  The scenes do sync back up if both players hit the reset button, so it can sort of work in that way, but it's not ideal since the plug essentially teleports on the observer's end because it places it in a different random location on each player's instance. The plug only syncs positions once the plug player moves it.