Code commenting, removing unused variables, and small bug fix.
While writing up some comments, I discovered a bug where on the observer view, when playing a second game using the restart button, the didWin boolean was initially never set back to false, so the win text would appear when the timer ran out. I added a line setting the boolean to false again in ResetGame, it was an easy fix.